remotefilelog: move getancestors into remotefilelog
Summary:
The getancestors call is a _very_ expansive at it requires walking all the
ancestors of the given node, which potentially requires O(N) network round
trips. Since we want to discourage/remove such behavior, let's move it out of
the store layer, and move it one layer up. The handful of places that calls
into ancestormap would need to be optimized to not call it later.

Reviewed By: DurhamG

Differential Revision: D17946852

fbshipit-source-id: 93eb7873b685ee88f8af5d4ceca500738d779396