remotefilelog: add remotefilelog.useruststore
Summary:
When set, the code will use the Rust ContentStore and MetadataStore
exclusively, bypassing very large portions of the Python code. In fact, the
only Python code that may run is when fetching data via SSH, the rest is in
Rust.

Things that are missing in Rust land that will are being tackled:
 - No memcache client,
 - Always uses SSH, not the new HTTP client,
 - Doesn't display stats about how much data was fetched after a command ran

When the option is turned on, the last point is responsible for most of the
remotefilelog test failures.

Reviewed By: quark-zju

Differential Revision: D18067661

fbshipit-source-id: b66b3fd0be2afa5114718a74d10f5f8958c3b34e