Properly parse empty diff-tree output
Summary:
A merge commit that merges parents with identical trees (e.g. `e964f3a095dd`
in bionic) will have a completely empty `diff-tree` output. Our `git diff-tree`
output parsing does not handle this case, as it currently uses two consecutive
newlines as a delimiter and recognizes that delimiter by looking for an empty
line from `readline()`. This recognition is flawed because it is not guaranteed
to consume two newlines: in the case where the output is empty, it will stop
after only one.

To fix this, change our delimiter to be just a single newline and call
`readline()` exactly once. This does not correctly handle filenames that
contain newlines, but neither did the previous method; I've added a comment
to make sure we address that in the future.

This is what was causing `diff-tree` to hang: because we wouldn't always read
its full output, eventually its stdout pipe would fill up and its `write()`
calls would block. Now that this is fixed, get rid of the restart hack we had.

Reviewed By: mwdevine

Differential Revision: D18319623

fbshipit-source-id: b24d55dd768145d5d27d6051b53aca90069068f8