changegroup: do not show "(+/- ? heads)"
Summary:
In an upcoming change we'll replace `cl.heads()` with `repo.heads()`.  In
changegroup code path the repo can be in an inconsistent state (ex. a bookmark
refers to an unknown commit) which breaks `repo.heads()`.

Since the whole purpose of the heads calculation is just to show
`(+/- ?  heads)`, which does not affect any real features, let's just
remove it.

Reviewed By: markbt

Differential Revision: D18366735

fbshipit-source-id: 893be2cec0c32b64a80b3ef4ca65b69f8ed76b27