mononoke: rework CoreContext into session and logging info
Summary:
This reworks our CoreContext a little bit to contain two fields instead of putting everything together in just one `Inner` struct. There's a few reasons why I'd like to make that change:

- First, with the `Inner` approach, everything has to be cloneable in `Inner`, even things that are largely static for a given session, because that's how we create new contexts to update their logger or Scuba sample. If we split things out, then we can clone the Logger & Scuba without cloning Inner.
- Second, this approach allows for better separation of concerns in the repo handler. Right now, it's a bit of a mess: many of our methods there aren't actually providing their command in Scuba for example, because they just use `self.ctx` and forget to update the method. It's just too error-prone. By separating things out, we have a data model that maps a little better to the state of the world (one session, multiple commands), and we make sure that we can't accidentally use a `ctx` without first tying it to a request. We had `prepared_ctx` as an attempt to do that, but since it wasn't mandatory it wasn't used everyhwere properly. The replacement `command_ctx` forces command code to pass their command in order to acquire a `CoreContext`.

Note that this diff is a lot of busywork here and there to update callsites accordingly. That said, there is one functional change in the commit cloud bookmarks filler, which was using a method it didn't need (and which never worked), so I took that out.

Relatedly, note that I removed `CoreContext` from the bundle2 parsing code, and passed just a `Logger` there. That code doesn't actually use `CoreContext`. This allowed for removing a few TODO's of dtolnay's when he introduced `FacebookInit`.

Reviewed By: StanislavGlebik

Differential Revision: D18352597

fbshipit-source-id: cd91042cef666c38b9cbd5f07518bc558e172aa2