mononoke: CoreContext: move PerfCounters to the command scope
Summary:
Right now, our PerfCounters live in the CoreContext's SessionContainer, so they are shared across all commands for a given session. One downside of that is that if a session has a lot of command, it's hard to:

- Figure out what the total value is for the PerfCounters (you have to look for the max on all the commands in the session — if you sum, you'll double-count a lot of things).
- Figure out what the value is for a PerfCounter for a given command (that one effectively impossible).

With this change, our PerfCounters are tied to individual commands, so we can figure out the total value by summing, and get the value for a given command directly.

Reviewed By: mitrandir77

Differential Revision: D18371382

fbshipit-source-id: 377a6594a95f47fcbed51361f4f457099c414962