mononoke: repo_client: eliminate duplication in Wireproto / command finished logging
Summary:
Currently, each individual command is responsible for logging its own `Command Processed` output. We typically want that to have perf counters (and wire proto logging), so that also has to be done on a per-command basis (and if we want to add new things to log when commands finish, then that would have to be done on a per-command basis as well), which results in quitea lot of code duplication / overlap.

This refactor reworks said command logging to eliminate this duplication by routing everything trough a single `CommandLogger` instance for each command (which is obtained along with a command's context).

This also removes a bit of the duplication we had around logging new commands (it's now in a `start_command` method. It also removes a now-useless with_logger_kv call that was adding the command to a logger that already has it.

There is still a requirement for each command to actually call `finalize_command` in order to log, but that's arguably still a step forward :)

Reviewed By: mitrandir77

Differential Revision: D18371381

fbshipit-source-id: a9bccb64120fee5c68633d3b43a8850416e2ffd4