mononoke: split uploaded_hg_bonsais_map into bonsais and hg_cs_ids
Summary:
In the previous refactorings I preserved the mapping between the originally
uploaded Mercurial hashes and the way Mononoke saved them in order to decide
whether we need it later: it seemed easier to keep it than to add it later.
But it does seem like the wrong thing to do: we do not need the mapping for
any purposes, we just need Mercurial changeset ids themselves to run hooks.
So let's separate concerns and preserve two different things:
`BonsaiChangeset`s to run `unbundle` and `HgChangesetId`s to run hooks.

Reviewed By: farnz

Differential Revision: D18421050

fbshipit-source-id: cd28e56465ae0d3d96381072de1f9bc5bb009516