Log to Scuba whenever pushredirection kicks in
Summary:
We're going to pushredirect some commits from small repos to large repos as part of the megarepo write path. Add some Scuba logging, so that we can see when redirection happens and react accordingly.

Note that I've deliberately kept the logging small - just tells you about the target repo - to avoid filling up Scuba. We can increase or reduce the amount of logging as we test this code.

Reviewed By: StanislavGlebik

Differential Revision: D18405345

fbshipit-source-id: bafc8f0aa0b4329b261dc0d6c99306fc9df95cf9