revert processing renames in file history
Summary:
[the diff is create with `hg backout`]

I've noticed slowness in Mononoke APIServer `get_file_history` due to recently deployed changes introduced in D18138431.
The slowness (that followed with timeouts) was caused by fetching a large changeset while checking renames for the paths, that were touched by the changeset.

Reverting the feature as it causes fallbacks while the rename history is not renderred by Diffusion anyway.

Reviewed By: StanislavGlebik

Differential Revision: D18430395

fbshipit-source-id: 2fbf9376d370624435f3846c7c88a1c5b9a53021