remotefilelog: add depth parameter to prefetchtrees
Summary: The top-level prefetchtrees function didn't support a depth parameter, while the underlying implementation does. Let's add an argument to pass on so that callers of the top-level function can configure the depth. The particular use case I have in mind is EdenFS's import helper, which needs to be able to limit the fetch depth to 1.

Reviewed By: quark-zju

Differential Revision: D18263661

fbshipit-source-id: 59991da2c30a5107df86c5a4b9614f113327f468