treemanifest: factor out HTTP fallback logic into new method
Summary: There are several places where we need to perform an HTTP tree fetch, and fall back to SSH if it fails. So far, every callsite has had to wrap the fetch function with a try block and manually print an error message and log the exception. Let's put all this functionality into a single wrapper function that catches all exceptions and returns a boolean indicating success or failure.

Reviewed By: quark-zju

Differential Revision: D18262915

fbshipit-source-id: b5c3395509eb90f70311f439129a6d60089bfe6b