sparse: use always matcher more consistently
Summary:
The `repo.sparsematch(*revs)` API is expected to return a union matcher that
unions sparse configurations from multiple revs.

I expect `repo.sparsematch(a, b, c)` to match more files than
`repo.sparsematch(a, b)`. However, that's not true if all sparse profiles in
`repo.sparsematch(a, b)` are empty (which, turns it into an always matcher),
and `repo.sparsematch(c)` is not empty.

Fix it by using the alwaysmatcher if any of the revs passed to `sparsematch`
has an empty profiler, instead of requiring all revs to all being empty
profiles.

Reviewed By: sfilipco

Differential Revision: D18414145

fbshipit-source-id: 357e90961acb9b2ae59fbdf84509f8201f629d44