pathmatcher: make TreeMatcher non-recursive
Summary:
Previously I thought it's possible to use `[pattern, "!%s/*" % pattern]` to
emulate non-recursive rules. However, that's not true.

Consider this rule `**/a`, if we translate it to `["**/a", "!**/a/*"]`, then
consider this path `a/b/a`:
- Its parent directory `a/b` matches the negative pattern `**/a/*` so the
  directory wouldn't be visited.
- If `a/b` is not visited, then `a/b/a` might be missed.
- But `a/b/a` does match the pattern `**/a`.

I couldn't find a way to express the non-recursive `**a` pattern using
recursive patterns.  Therefore,  change TreeMatcher APIs to do non-recursive by
default. The callsite can still append `/**` to all rules to get the old
behavior.

Reviewed By: sfilipco

Differential Revision: D18342228

fbshipit-source-id: a97144c2a73c6d35fbfbd0d55ea3661ad4a9ac12