mononoke/repo_client: add support for commit rate limiting
Summary:
This adds support for limiting the number of commits for a given author. This
runs after we've received Bonsais from the client, but before we attempt to
pushrebase them or anything. This is all controlled through Configerator,
through the same configuration as our throttling.

I'll also send a diff in Landcastle in order to make this non-retryable.

Differential Revision: D18375115

fbshipit-source-id: 089bdcd7bebfd2ea42c37921fc80b53f96a1d40e