Implement file mode logic for repo converter
Summary: Up until now, we haven't been setting file modes for the files we're putting into hg. This seems like it might be related to some of the problems we're seeing in changes relating to symlinks. This commit adds mode tracking to try to fix that. It also adds unit tests to test the underlying mode conversion functions in gitutil to check our logic.

Reviewed By: tchebb

Differential Revision: D18409701

fbshipit-source-id: c5c3a90e12bd7115e24024a912f175484efe8a90