zstore: blob store backed by zstdelta
Summary:
This blob store is the first usage of indexedlog. It uses zstd dictionary
compression provided by the zstdelta crate and aims to replace lz4 + bdiff
revlog for storing file revisions.

Blobs are enforced SHA1 addressed. A separate mapping from filenode to SHA1 is
needed if we want to integrate it into Mercurial file storage.

Practically, compare with lz4+bdiff, this setup compresses better (which could
be improved by tweaking parameters in zstdelta - it uses a variant of zstd
level 1 currently) when using same delta-chains:

| Repo    | File revisions | Zstore size (log + index) | Revlog du size | Revlog raw size |
| hg      | 74.9K          | 28 + 2.1MB                | 40MB           | 33.3MB          |
| cpython | 232K           | 179 + 6.5MB               | 221MB          | 199.9MB         |

It should also compress better for binary files, since there is no line-based
diff algorithm.

Performance wise, my previous experiments on fbsource showed compression is
faster with zstdelta. Decompression is slower. But in reality we might have
many files to decompress, in which case, multi-thread should make things
better.

The delta base decision logic is designed to automatically reshape a linear
chain to a tree to reduce chain length. The next diff adds dedicated tests
for the algorithm.

The name `Zstore` was chosen because the prefix `zst` is the name of
the compression algorithm.

Reviewed By: kulshrax

Differential Revision: D10251983

fbshipit-source-id: 3ba308d2799bb759a36f69a88f7c0144830fc41e