mononoke/integration_runner_real: make the mode/opt check more robust
Summary:
StanislavGlebik reported that running the tests locally in mode/opt was failing. The
reason for that is because while `buck test` sets `FBCODE_BUILD_MODE`, it's not
set when the binary runs.

This fixes that by looking up our manifest to know if our binary is optimized
or not... which is kinda clowny but it will have to do.

Reviewed By: farnz

Differential Revision: D18507748

fbshipit-source-id: ed6f545df61a3f3be050e7cc519f6a4e39e2e4aa