rust/failure_ext: Support 1 or 2 type params in Result
Summary:
The new type alias allows being used as both `Result<T>` and `Result<T, SomeOtherError>`. This is useful when almost every Result in a module uses failure::Error as the error type but there are a few special-purpose signatures with a different error type; we still want to import failure_ext::Result, but keep using Result<T, Other> where needed.

This is forward-compatible with migrating to https://docs.rs/anyhow/1.0.19/anyhow/type.Result.html which follows the same scheme.

Reviewed By: Imxset21

Differential Revision: D18499669

fbshipit-source-id: e9ad9b2853fc0a16c56018573f9e64fbee363f61