common/rust: Remove E argument of with_context closure
Summary:
As evidenced by the fact that it is not used anywhere, this argument is not useful.

Before:

```
.with_context(|_| format!(...))
```

After:

```
.with_context(|| format!(...))
```

In the case that someone in the future does need a reference to the error value in a context, the recommended way to write would be:

```
.map_err(|e| {
    /* whatever you want to do with &e */

    e.context(format!(...))
})
```

This diff is in preparation for migrating off of failure::Fail to std::error::Error for error types, in combination with anyhow::Error whose [with_context closure does not take an argument](https://docs.rs/anyhow/1.0.19/anyhow/trait.Context.html).

Reviewed By: Imxset21

Differential Revision: D18498898

fbshipit-source-id: 2c6f1929944d5877c408e367cde3c99ab2a5a07a