inlude copy/move information in commit_compare API
Summary:
D18224367 introduces `compare_with_parents` that also includes copy/move information but it turned out that we'd also need to do manifest diff in that case as the bonsai doesn't have all the necessary info (so the perf of that call becomes very similar to this one).

Until I figure out what I want to do with the other call I can still provide
some copy/move information so we can start redirecting diff traffic to the new service.

Reviewed By: markbt

Differential Revision: D18505946

fbshipit-source-id: 12bebb368f999afff01d3a5f8aebc97f338ff162