cat command for fetching file contents
Summary:
Add an `scsc cat` command which can be used to fetch file contents at a
particular revision.

This uses chunking to ensure the server doesn't need to buffer the whole file
in memory, and to provide some streaming of the output.

JSON output is less useful for `cat`, but it is still provided.  File chunks
that are falid UTF-8 are rendered as strings.  File chunks that are not valid
UTF-8 are hex-encoded.

Reviewed By: ikostia

Differential Revision: D18528708

fbshipit-source-id: 38f82d05d956928fe655e7cf9862e7c0a57d97ac