mononoke: make get_hg_bonsai_mapping less surprising
Summary:
get_hg_bonsai_mapping was very easy to misuse. If you pass a list of bonsai
changesets then it just searched for corresponding hg changesets at the
bonsai_hg_mapping database.

The problem is that hg changesets might not be generated yet for bonsai
changesets, and normally the callers would expect that hg changesets will be
lazily generated. In apiserver we had a workaround for that, however it seems
better to move the logic in the blobrepo itself.

A few notes:
1) We don't need to change anything for hg changesets, because bonsai commits
shouldn't be generated from hg changesets.
2) Note that to preserve semantic of this function we don't fail if bonsai
changeset doesn't exist.

Reviewed By: krallin

Differential Revision: D18394642

fbshipit-source-id: 977aab08f3321b9797fa8c032f1d83c0ef7ff487