mononoke/lfs_server: make throttling more generic
Summary:
This builds on my earlier diff to add throttling to the download endpoint, but
makes it more generic to allow for throttling to be defined in terms of any ODS
counters, dynamically in live configuration (i.e Configerator).

This will give us flexibility not only to adjust thresholds in real-time, but
also tune what we are rate-limiting on (we do not, however, have matching on
what the request is, so it's still a fairly blunt instrument).

This is implemented as Gotham middleware since we want this to be able to
provide a response (and do so asynchronously), and we only care to run it if
we're not going to return a 404 or 405.

Reviewed By: farnz

Differential Revision: D18478276

fbshipit-source-id: ddce97380ec9e44f2350e6be61ffc277cd7b738d