mononoke/lfs_server: add throttling support
Summary:
This adds throttling suppot in the Mononoke LFS server. This will let us drop
some requests if we're overloaded, which is generally desirable to make sure we
can recover if e.g. Sandcastle has an outage and they need a _lot_ of data from
us.

I implemented this by reading our own counters (see the comments in the diff
for rationale, and the next diff in this stack), so I also refactored our stats
aggregation to always run even if fb303 is not enabled (which we weren't doing
before).

Reviewed By: ahornby

Differential Revision: D18478277

fbshipit-source-id: a88c2a47446ff65bf83a31485d38859ee92eb5e8