mononoke/lfs_server: don't use presence cache
Summary:
TL;DR: This should shave off ~40ms from single-blob batch requests to the LFS
Server (mostly Eden AFAIK). Which is about approximately 100% of the time those
requests take.

When a query hits the presence cache in cachelib and misses, the presence
response from upstream doesn't get cached, so we end up always going to
Memcache (or even the blobstore if this isn't cached in Memcache either, which
we only fill if some other host does a get). This query updates the LFS Server
to just try to fetch the object instead, which we cache properly.

The upshot of this change is that on a hot cache, this results in the floor for
batch requests being around 140us instead of 40ms (that's a 285x performance
improvement!). For the time being, this is particularly important for Eden,
which does **a lot** of 1-object batch requests
(https://fburl.com/scuba/yk4ew6um)

As we'll see below, the cache seems likely to be hot, because those things are
very small and tend to be very hit (but I will definitely double-check that
after this ships).

There is the caveat of loading the objects instead of just their presence
bits. As it turns out, that's OK: the contents we are caching here are the file
"indices" (we're talking large files here), which are quite small.

Indeed, a review of hits from the LFS server to load such files (we download
them when serving the files) shows that they are 440B on average
(https://fburl.com/scuba/5tweeulh), and we serve about 8000 distinct files
total (https://fburl.com/daiquery/u8e12w8a), which means we're talking about
3.5 MiB of cache usage.

One other concern might be that these file pointers are now contending with
file data for cache space, so they could get evicted from cachelib. Obviously,
that's still better than not being in the (presence) cache at all, but still.
However, note that we already have this "problem" because forward pointers are
blobs too, which means they contend with files as well (in other words, we're
making this "problem" potentially twice as bad, but not materially changing
it).

We could (and should) improve that further by storing deserialized file bytes
in a separate cache altogether instead of relying on the same blob cache for
those.  That said, this isn't a *huge* issue in practice. Indeed, note that
even before we introduced consistent routing, the P95 headers duration on
download (which does a look up for the exact same blobs this diff looks for)
was around 0 (meaning it hit in cachelib): https://fburl.com/scuba/0i0rpmmk

In other words, there seems to be good reason to believe that we'll see cache
hits for 95% of requests. If we wanted to cheat, we could also update hg to set
`?routing=` when it's about to make a batch request for just one object (and
then we'd expect 100% cache hits).

Longer term, I'd like to overhaul our caching a little bit to make presence
checks fast and isolate blob caching from file content caching in the LFS
Server, but for now, let's not let the better be the enemy of the good, and
let's take the (anticipated) (massive) perf improvement.

Reviewed By: StanislavGlebik

Differential Revision: D18572151

fbshipit-source-id: dfffa98f25a8045c074e69aa4dd49c843a663e48