mononoke/lfs_server: publish PerfCounters for each request
Summary:
This updates our LFS Server to publish PerfCounters to Scuba on a per-request
basis. We've been wondering what our cache hit rate on a per-request basis so
this will be a nice way to answer this.

A separate goal is also to perhaps eventually convert the API Server to use the
same HTTP stack as the LFS Server, which will let us access cache hit rate for
the API Server on a per-HTTP-request basis.

Reviewed By: farnz

Differential Revision: D18505633

fbshipit-source-id: fdd27ffdd901a21527cf5de1dd8fd27c9a94e334