mononoke/cacheblob: track cache hits and misses
Summary:
This updates cacheblob to track cache hits and misses in CoreContext perf
counters for Cachelib and Memcache. It also allows any CacheOps struct to
provide its own counters to track.

Note that as part of this I've had to make the types concrete in Sqblob, since
we otherwise cannot use associated types.

Reviewed By: farnz

Differential Revision: D18505170

fbshipit-source-id: 6241f4ca0e8b8cd58427b79869f4d67cc47961dd