mononoke: lfs: use X509 cert to verify proxy identity
Summary:
Previously, we were reading forwarded client identities from the
HTTP header. We should only do this if we have verified that it is a trusted
proxy giving us this header.

Update the LFS server to pass the identities from the X509 cert to the client
identity middleware, such that we parse the HTTP header if the client identity is
the trusted proxy else we take the identities from the X509 cert.

Reviewed By: krallin

Differential Revision: D18452652

fbshipit-source-id: c5d06bd07ee84864413a59b5e391ef9fd00819a0