mononoke: Add AclChecker to LFS Server
Summary:
Now that the LFS Server can correctly verify identities, AclChecker can be
used to ensure that only authorised users can interact with the LFS server.

If the acl_check config option is set to False, then this change is a NOP.

If the hipster_acl config option is set, then a new AclChecker will be created that will
be queried after each connection is established. If it isn't set, then all clients with
identities will be allowed to connect - if a client doesn't provide an ident then it
will be rejected.

NOTE: Acl checking is disabled by default.

Reviewed By: krallin

Differential Revision: D18452653

fbshipit-source-id: 631b34f70ee074c0a7d53c0c4fef4ea27bf13a2c