mononoke/lfs_server: add test-friendly logging
Summary:
StanislavGlebik reported that some of the LFS tests were occasionally flaking out. This
is because we asynchronously resolve the hostname of the connecting client
before logging, and if the client is making requests fast enough we might
occasionally end up with logs out of order.

This fixes the problem by adding a test-friendly log mode that logs
synchronously and doesn't wait for hostnames. This is also nicer when writing
tests since it means we can just let the integration runner copy output and
don't need any globs.

Reviewed By: HarveyHunt

Differential Revision: D18592900

fbshipit-source-id: 1bafd684d8fc710dfb09cb8453bb16de774ad02d