mononoke: check pushredirector status in backsyncer_cmd
Summary:
We'd like to have backsyncer_cmd report how many pushes are left to backsync so
that we can add an alarm on that. But this alarm should also work fine with
frequent rollout and rollbacks that we are going to make.

This diff changes `backsync_forever` function to read configerator killswitch
and do not import anything if killswitch is disabled. This diff also changes
ods reporting - it now reports 0 if push redirection is disabled. That means we
can add alarm on the queue size and make sure that it won't misfire when push
redirection is disabled.

This diff also did a few refactorings:
1) backsync_forever was only used in backsyncer_cmd, so it was moved there
2) backsync_forever now just calls backsync_all_latest to avoid code
duplication

Reviewed By: farnz

Differential Revision: D18617272

fbshipit-source-id: 49cdc8211c97c7d8473f016835e4e55916209b71