mononoke: report bookmark age difference from mononoke_api
Summary:
This diff introduces an ability for `mononoke_api` to report stats about
it's health.
In particular, one implemented stat is a differences in age for a configured
set of bookmarks. It is calculated as the difference of timestamps for
`author_date` field in the `BonsaiChangeset`, acquired via the normal
`mononoke_api` flow vs one acquired from a blobrepo (hopefully guaranteed to
not be stale).

The idea is to allow the users of `mononoke_api` to call this periodically. The
use in `scs_server` is to come in the following diff.

Reviewed By: markbt

Differential Revision: D18550868

fbshipit-source-id: ff75f76cd8a20de9371cb32d6a057dca7c18f04d