mononoke: Fix flaky lfs_server tests
Summary:
A couple of LFS server tests are flaky due to the expected output not
appearing in the scuba log file. The affected tests run curl to send a command
to the LFS server. A theory for the missing output is that curl completes so
quickly that the async write to the log file hasn't happened before the test
checks the contents of the log file.

Fix this by adding a test helper function that waits until a file is non empty.
Further, update the scuba logging code to seek to the end of the file before
writing so that it doesn't create sparse files if the log file is truncated.

Reviewed By: krallin

Differential Revision: D18687831

fbshipit-source-id: 46b93306a3ac694064ab5cd4acd5eff25ab320e3