mononoke: derive deleted files manifest
Summary:
Implementation of derive method for Deleted Files Manifest, which receives bonsai changeset id, changes associated with the changeset and list of parent deleted manifests and constructs root manifest for the current changeset.
This implementation only works with linear histories, only with commits with single parent. Support for merges is coming soon.

Reviewed By: StanislavGlebik

Differential Revision: D18204686

fbshipit-source-id: f30c4f045109db0470d8936c26f2f6ef986b87f8