mononoke/lfs_server: add support for capturing TLS Session Data
Summary:
This adds support in the LFS Server for logging TLS Session Data. Effectively,
this makes it possible to decrypt traffic to Mononoke's LFS Server using
tcpdump and wireshark.

To do so, we log the TLS master secret in a file using [Mozilla's key log
format](https://developer.mozilla.org/en-US/docs/Mozilla/Projects/NSS/Key_Log_Format),
which Wireshark understands.

This is all but intended for daily use: the end goal is exclusively to assist
with debugging to understand the 40ms delay we're seeing reading the body on
batch requests. The goal isn't to run this all the time (!).

Instead, the workflow would be to kick off a canary with this, collect a bit of
traffic, then turn it off.

Reviewed By: farnz

Differential Revision: D18689109

fbshipit-source-id: 86d626f9acd37f632bf1e7a0bdd2b967d10edfc4