Split SCS thrift service trait from the main implementation
Summary:
`SourceControlServiceImpl` has duplicate methods for each of the thrift
methods: one base method that implements the functionality, and one trait
method that delegates to the base method.

This could be confusing if we fail to resolve the correct method.

Instead, split the two out to separate structs by making the thrift server a
wrapper around the base implementation that only contains the delegation
methods.

Reviewed By: mitrandir77

Differential Revision: D18745443

fbshipit-source-id: 5acdad2f1f66562ecf71f85ebf2b83d15f0edc9a