Split up SCS trait implementation
Summary:
The trait implementation for the source control service is also too large, and
will only grow larger as we implement more methods.

It's not possible in Rust to split a trait implementation into multiple files.
Instead, we implement the methods directly on the `SourceControlServiceImpl`
struct, and then use a macro to generate delegation methods in the same way
that the `async_trait` proc macro would do.  This lets us put each of the
implementations in different files.

Reviewed By: mitrandir77

Differential Revision: D18745178

fbshipit-source-id: ae088b6d059b3dff0ee28563c33fada0900dd640