mononoke: address followup from S191123
Summary:
This diff makes sure that filenodes are generated when merges are pushed. See
comments in pushrebase/src/lib.rs for more details.

Overall this is a hack and it's caused by the fact that we want to generate
filenodes only for public commits, but there's no easy way to tell if commit
will become public at the point when we generate hg changesets.
I think that the correct solution we should aim to is generating completely
filenodes outside of hg changeset generation. T58398518 has more details.

Reviewed By: farnz

Differential Revision: D18748102

fbshipit-source-id: 7d15d5399315464235c865e77f5fd4030c109c41