blobstore_healer: Fix downcast to mysql_async::Error
Summary:
Changing common/rust/sql's errors from failure::Fail trait to std::error::Error trait broke a downcast in Mononoke -- since mysql_async 0.20 does not provide a std::error::Error impl for its error type, we had been using failure_ext::convert to convert those errors from `impl Fail` to `impl Error`. This conversion preserved the cause chains but did not preserve the concrete type of the error as mysql_async::error::Error.

This diff adds a check for the specific variant of mysql_async::error::Error that we care about downcasting to, prior to the failure_ext::convert conversion.

The real fix will be to upgrade to mysql_async 0.21 as soon as possible, as that release provides a correct std::error::Error impl for the error type. This workaround can be reverted when that upgrade is done.

Reviewed By: farnz

Differential Revision: D18789215

fbshipit-source-id: 934943e5fe93fbd942143a496278d74cfa294d2b