eden: redirect: use HFS+ rather than APFS
Summary:
We've seen occasional weirdness with "resource temporarily unavailable" when working with disk images.

The internet, in the form of this stackoverflow post: https://stackoverflow.com/q/48368389/149111
suggests that this is related to APFS.

This commit switches over to HFS+ in the hope that we won't see it any more.

Reviewed By: chadaustin

Differential Revision: D18795799

fbshipit-source-id: 68e08e852770e311bcf04a8d12cb20670babf889